(refactor) Moves generic utility method to generic utilities
