+++ agents/utils/actions.py
@@ -71,18 +71,6 @@ class JointsData:
         return mapped
 
 
-def find_missing_values(check_list, string_list: List) -> List:
-    """
-    Return strings from `string_list` that do NOT appear in the dictionary's values.
-    """
-    # Convert values to a set for efficient lookup
-    value_set = set(check_list)
-    # Collect strings that are not found among the values
-    missing = [s for s in string_list if s not in value_set]
-
-    return missing
-
-
 # Map requirement categories -> URDF limit keys
 req_map = {
     "positions": ["lower", "upper"],
